Update README.md
Our App provides solutions to current time Energy. There is a need for renewable energy generation the main reason is to reduce carbon emissions and produce energy eco-friendly. Our App provides a solution for Renewable Energy investors, who are interested to set up renewable energy generators. Many investors face the problem of correctly identifying the location and other parameters which are involved in energy production. Also, investors are interested in the Return on Investment. Our App solves both these problems and provides a stepwise selection of location and other details. Then based on the weather information available at the specific location.

Our mobile application, Stellar, brings forward NASA's data. This app is mainly for investors who are looking to invest in solar panel brands or a certain duration. Users can choose the specific city that they want to select as the solar panel and can also choose the area of the panel, efficiency, and many more features. As the program opens, there are 3 interesting facts depicted regarding solar energy and hoow it's good for the people. Upon clicking "getting started", the user is taken to the apge where they get to select from the clean energy=producing cities. After this, the user is asked which energy generation infrastructure to use as well as how much amount they want to invest. Upon submitting the amount, the user will be shown a drop down menu of a list of brands of solar panels to choose from and will have the neter the area required to setup the plant. Also, the user is asked the duration they want to invested their money for. Lastly, the user is shown the final cost of setting up the plant and the Return on Investment (ROI) and the power generation is displayed. In the app, users can find the ROi and can also find the power generated. OUr app is useful for investors looking to put money into solar panel brands and analyze the different features that the solar panels comprise of. We used Python, Flutter, and Dart, for the design and making of the front end and the back end of our application

The Space Apps Experience was a very unique experience and it will be a great learning curve for us, we have gained a lot of knowledge through research and practising code on a high level. Skills such as data gathering and data cleaning were heavily looked into, we all learned how important it is to locate data and be consistent with our data. Moreover, calculations for the energy and the power output was a key for us and we had to integrate those calculations along with the dataset into Python. We used the NASA's Power API as well as OpenWeather API for additional functionality in our application.

NASA space apps team: https://2022.spaceappschallenge.org/challenges/2022-challenges/take-flight/teams/idealbits/project